+++ AdoptaPatas/src/main/java/org/programarteduele/app/repository/CommentLikeRepository.java
@@ -4,7 +4,7 @@ import org.programarteduele.app.entity.CommentLike;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface CommentLikeRepository extends CrudRepository<CommentLike, Integer>{
+public interface CommentLikeRepository extends CrudRepository<CommentLike, Long>{
       
 
 }
