+++ AdoptaPatas/src/main/java/org/programarteduele/app/repository/PostLikeRepository.java
@@ -4,7 +4,7 @@ import org.programarteduele.app.entity.PostLike;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface PostLikeRepository extends CrudRepository<PostLike, Integer>{
+public interface PostLikeRepository extends CrudRepository<PostLike, Long>{
       
 
 }
